QC dashboard: lift scrnaseq extractor + summary into the template module
Moves _extract_metrics, _read_*, _generate_summary, _collect_plots,
_get_results_bucket from QCDashboardService into:

- app/services/qc/templates/scrnaseq.py -- extract(), generate_summary(),
  read_starsolo_summary(), read_h5ad_metrics(), read_multiqc_metrics(),
  read_multiqc_chart_data(), build_barcode_rank_data(),
  extract_barcode_rank_from_mtx(), read_umi_per_cell_sorted(), and the
  MULTIQC_PLOTS list. Logic is verbatim -- existing scRNA-seq runs rate
  and summarize identically.
- app/services/qc/extractors/gcs_helpers.py -- get_results_bucket()

QCDashboardService is now a thin orchestrator: resolves template, asks the
template for metrics + quality + summary, persists with the snapshotted
config. Backwards-compat shims (_read_starsolo_summary, _compute_quality_rating
etc.) stay on the class so test_qc_plot_upload.py continues to pass against
its existing mock surface.

Custom pipelines: dispatcher reads qc_metrics.json from the run prefix, no
template-side extraction work needed -- the pipeline owns its own emission.
